# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Eco Escuela
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ObraStatus = typeof ObraStatus[keyof typeof ObraStatus];


export const ObraStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Obra {
  id: string;
  localId: string;
  studentName: string;
  title: string;
  description: string;
  category: string;
  hours: number;
  date: string;
  photo?: string | null;
  status: ObraStatus;
  /** @nullable */
  teacherComment?: string | null;
  createdAt: string;
  /** @nullable */
  reviewedAt?: string | null;
}

export interface ObraInput {
  localId: string;
  studentName: string;
  title: string;
  description: string;
  category: string;
  hours: number;
  date: string;
  photo?: string | null;
}

export type ObraReviewStatus = typeof ObraReviewStatus[keyof typeof ObraReviewStatus];


export const ObraReviewStatus = {
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface ObraReview {
  status: ObraReviewStatus;
  /** @nullable */
  teacherComment?: string | null;
}

export interface TeacherTokenInput {
  token: string;
}

export interface SimpleOk {
  ok: boolean;
}

export type ListObrasParams = {
studentName?: string;
status?: ListObrasStatus;
};

export type ListObrasStatus = typeof ListObrasStatus[keyof typeof ListObrasStatus];


export const ListObrasStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;