# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Eco Escuela
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: obras
    description: Obras ecológicas operations
  - name: notifications
    description: Push notification token management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /obras:
    get:
      operationId: listObras
      tags: [obras]
      summary: List all obras
      parameters:
        - in: query
          name: studentName
          schema: { type: string }
          required: false
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
          required: false
      responses:
        "200":
          description: List of obras
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Obra"
    post:
      operationId: submitObra
      tags: [obras]
      summary: Submit an obra ecológica
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObraInput"
      responses:
        "201":
          description: Obra created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obra"

  /obras/{id}:
    get:
      operationId: getObra
      tags: [obras]
      summary: Get a single obra
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Obra found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obra"
        "404":
          description: Not found

  /obras/{id}/review:
    patch:
      operationId: reviewObra
      tags: [obras]
      summary: Teacher reviews an obra
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObraReview"
      responses:
        "200":
          description: Obra reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obra"
        "404":
          description: Not found

  /teacher-tokens:
    post:
      operationId: setTeacherToken
      tags: [notifications]
      summary: Register a teacher push notification token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherTokenInput"
      responses:
        "200":
          description: Token registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleOk"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status: { type: string }
      required: [status]

    Obra:
      type: object
      properties:
        id: { type: string }
        localId: { type: string }
        studentName: { type: string }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        hours: { type: number }
        date: { type: string }
        photo: { type: string, nullable: true }
        status:
          type: string
          enum: [pending, approved, rejected]
        teacherComment:
          type: ["string", "null"]
        createdAt: { type: string }
        reviewedAt: { type: ["string", "null"] }
      required: [id, localId, studentName, title, description, category, hours, date, status, createdAt]

    ObraInput:
      type: object
      properties:
        localId: { type: string }
        studentName: { type: string }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        hours: { type: number }
        date: { type: string }
        photo: { type: string, nullable: true }
      required: [localId, studentName, title, description, category, hours, date]

    ObraReview:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
        teacherComment:
          type: ["string", "null"]
      required: [status]

    TeacherTokenInput:
      type: object
      properties:
        token: { type: string }
      required: [token]

    SimpleOk:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]