# lib/api-zod/src/generated/types/obra.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Eco Escuela
 * OpenAPI spec version: 0.1.0
 */
import type { ObraStatus } from './obraStatus';

export interface Obra {
  id: string;
  localId: string;
  studentName: string;
  title: string;
  description: string;
  category: string;
  hours: number;
  date: string;
  photo?: string | null;
  status: ObraStatus;
  /** @nullable */
  teacherComment?: string | null;
  createdAt: string;
  /** @nullable */
  reviewedAt?: string | null;
}