# lib/api-zod/src/generated/types/obraInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Eco Escuela
 * OpenAPI spec version: 0.1.0
 */

export interface ObraInput {
  localId: string;
  studentName: string;
  title: string;
  description: string;
  category: string;
  hours: number;
  date: string;
  photo?: string | null;
}